Only display view all button if unshown doc aggs exist
