+++ context/app/static/js/components/entity-search/facets/select/SelectFacet/SelectFacet.jsx
@@ -1,32 +1,14 @@
-import React, { useState } from 'react';
-import Button from '@material-ui/core/Button';
+import React from 'react';
 import { useSearchkit } from '@searchkit/client';
 
 import SelectFacetItem from 'js/components/entity-search/facets/select/SelectFacetItem';
-import { useStore } from 'js/components/entity-search/SearchWrapper/store';
-import { defaultSelectFacetSize } from 'js/components/entity-search/SearchWrapper/utils';
+import SelectFacetViewButton from 'js/components/entity-search/facets/select/SelectFacetViewButton';
+
 import { StyledFormGroup } from './style';
 
-const sharedButtonProps = {
-  variant: 'text',
-  size: 'small',
-  color: 'primary',
-};
-
-function SelectFacet({ facet: { entries, identifier } }) {
+function SelectFacet({ facet: { entries, identifier, sumOtherDocCount } }) {
   const api = useSearchkit();
-  const { setFacetSize } = useStore();
-  const [isViewingAll, setIsViewingAll] = useState(false);
 
-  function handleViewAll() {
-    setFacetSize({ identifier, size: 10000 });
-    setIsViewingAll(true);
-  }
-
-  function handleViewLess() {
-    setFacetSize({ identifier, size: defaultSelectFacetSize });
-    setIsViewingAll(false);
-  }
   return (
     <>
       <StyledFormGroup>
@@ -40,15 +22,7 @@ function SelectFacet({ facet: { entries, identifier } }) {
           />
         ))}
       </StyledFormGroup>
-      {isViewingAll ? (
-        <Button {...sharedButtonProps} onClick={handleViewLess}>
-          View less
-        </Button>
-      ) : (
-        <Button {...sharedButtonProps} onClick={handleViewAll}>
-          View all
-        </Button>
-      )}
+      <SelectFacetViewButton identifier={identifier} sumOtherDocCount={sumOtherDocCount} />
     </>
   );
 }
