+++ context/app/static/js/components/entity-search/facets/select/SelectFacetViewButton/index.js
@@ -1,0 +1,3 @@
+import SelectFacetViewButton from './SelectFacetViewButton';
+
+export default SelectFacetViewButton;
